FragmentOffset获取和设置问题修复

+++ WindivertDotnet/FragmentFlag.cs
@@ -6,7 +6,7 @@ namespace WindivertDotnet
     /// IPv4标记位
     /// </summary>
     [Flags]
-    public enum FragmentFlag : ushort
+    public enum FragmentFlag : byte
     { 
         /// <summary>
         /// 有分片
